refactor(hooks): replace hooks to ng-cli fileReplacements logic
