+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ElectronService } from './providers/electron.service';
 import { TranslateService } from '@ngx-translate/core';
-import { AppConfig } from './app.config';
+import { AppConfig } from '../environments/environment';
 
 @Component({
   selector: 'app-root',
