+++ src/environments/environment.dev.ts
@@ -3,7 +3,7 @@
 // `ng build --env=prod` then `index.prod.ts` will be used instead.
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
-export const CONF_DEV = {
+export const AppConfig = {
   production: false,
   environment: 'DEV'
 };
