+++ src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-export const CONF_PROD = {
+export const AppConfig = {
   production: true,
   environment: 'PROD'
 };
