+++ src/main.ts
@@ -2,7 +2,7 @@ import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
-import { AppConfig } from './app/app.config';
+import { AppConfig } from './environments/environment';
 
 if (AppConfig.production) {
   enableProdMode();
